Do not abuse PathBuf::push("") for appending directory separator
Since https://github.com/rust-lang/rust/pull/89270 rust's pathbuf
push() normalizes the path if it is a verbatim path, so that the result
is still a valid verbatim path. This involves parsing reconstructing the
path.

Appending a path separator using push("") will no longer work, and if
it did, it would be very inefficient.

https://github.com/rust-lang/rust/issues/89658

Signed-off-by: Sean Young <sean@mess.org>